requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
streamlit==1.29.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
openai==1.3.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.3
numpy==1.26.2 